+++ src/main/java/com/tavemakers/surf/domain/activity/dto/response/ActivityPenaltyGroupReqDTO.java
@@ -1,6 +1,5 @@
-package com.tavemakers.surf.domain.activity.dto.test;
+package com.tavemakers.surf.domain.activity.dto.response;
 
-import com.tavemakers.surf.domain.activity.dto.response.ActivityTypeGroupCountResDTO;
 import lombok.Builder;
 
 @Builder
