+++ src/main/java/com/tavemakers/surf/domain/activity/mapper/ActivityRecordMapper.java
@@ -3,8 +3,8 @@ package com.tavemakers.surf.domain.activity.mapper;
 import com.tavemakers.surf.domain.activity.dto.response.ActivityRecordSummaryResDTO;
 import com.tavemakers.surf.domain.activity.dto.response.ActivityTypeCountResDTO;
 import com.tavemakers.surf.domain.activity.dto.response.ActivityTypeGroupCountResDTO;
-import com.tavemakers.surf.domain.activity.dto.test.ActivityPenaltyGroupReqDTO;
-import com.tavemakers.surf.domain.activity.dto.test.ActivityRewardGroupReqDTO;
+import com.tavemakers.surf.domain.activity.dto.response.ActivityPenaltyGroupReqDTO;
+import com.tavemakers.surf.domain.activity.dto.response.ActivityRewardGroupReqDTO;
 import com.tavemakers.surf.domain.activity.entity.ActivityRecord;
 import com.tavemakers.surf.domain.activity.entity.enums.ActivityType;
 import org.springframework.stereotype.Component;
